fix: scores 의 원소 타입이 int 였던 오류 수정

+++ src/perception/include/perception/Model.h
@@ -54,7 +54,7 @@ namespace layer_names {
 };
 
 struct DetectionInfo {
-    std::vector<int> scores;
+    std::vector<float> scores;
     std::vector<xt::xarray<float>> poly4s;
     std::vector<xt::xarray<float>> tri_ptss;
 };
